+++ WindowUtility.cpp
@@ -1229,7 +1229,7 @@ bool GetOpenFileName(
     ofn.lpstrFilter = ToWChar(fileTypes);
     ofn.nFilterIndex = 0;
     ofn.lpstrFile = ToWChar(filePathBuffer);
-    ofn.nMaxFile = std::size(filePathBuffer);
+    ofn.nMaxFile = uint32_t(std::size(filePathBuffer));
     ofn.lpstrTitle = ToWChar(title);
     ofn.Flags = OFN_FILEMUSTEXIST | OFN_HIDEREADONLY | OFN_NOTESTFILECREATE | OFN_EXPLORER | OFN_ENABLESIZING;
 
@@ -1267,7 +1267,7 @@ bool GetSaveFileName(
     ofn.lpstrFilter = ToWChar(fileTypes);
     ofn.nFilterIndex = 0;
     ofn.lpstrFile = ToWChar(filePathBuffer);
-    ofn.nMaxFile = std::size(filePathBuffer);
+    ofn.nMaxFile = uint32_t(std::size(filePathBuffer));
     ofn.lpstrDefExt = ToWChar(defaultExtension);
     ofn.lpstrTitle = ToWChar(title);
     ofn.Flags = OFN_OVERWRITEPROMPT | OFN_HIDEREADONLY | OFN_NOTESTFILECREATE | OFN_EXPLORER | OFN_ENABLESIZING;
